Seperated the parts of the pipeline to different files.
