+++ crates/bld_runner/src/lib.rs
@@ -3,6 +3,6 @@ mod platform;
 mod sync;
 
 pub use sync::builder::RunnerBuilder;
-pub use sync::pipeline::External;
-pub use sync::pipeline::Pipeline;
+pub use pipeline::Pipeline;
+pub use pipeline::external::External;
 pub use sync::runner::Runner;
