+++ crates/bld_runner/src/pipeline/mod.rs
@@ -1,2 +1,10 @@
 pub mod syntax;
 pub mod traits;
+pub mod runs_on;
+pub mod variable;
+pub mod external;
+pub mod step;
+pub mod artifacts;
+mod pip;
+
+pub use pip::*;
