+++ crates/bld_runner/src/pipeline/variable.rs
@@ -1,0 +1,14 @@
+#[derive(Debug)]
+pub struct Variable {
+    pub name: String,
+    pub default_value: String,
+}
+
+impl Variable {
+    pub fn new(name: String, default_value: String) -> Self {
+        Variable {
+            name,
+            default_value,
+        }
+    }
+}
