+++ crates/bld_runner/src/sync/mod.rs
@@ -1,3 +1,2 @@
 pub mod builder;
-pub mod pipeline;
 pub mod runner;
